+++ example/lib/main.dart
@@ -7,95 +7,196 @@
  */
 
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
 import 'package:memory_and_json_directories/memory_and_json_directories.dart';
 
-void main() {
-  String typeName = MAJDirectory().getTypeName();
+/*
+  The custom class used to build the CustomItemWidget
+  I recommend not having any constructors for this class
+  You should use the MAJNode.data property to pass any
+  data you would need in a constructor
+ */
+class CustomItem implements MAJItemInterface {
+  static const String typeName = "custom_item";
 
+  @override
+  String getTypeName() {
+    return typeName;
+  }
+
+  @override
+  Widget build({
+    required BuildContext context,
+    required MAJNode nodeReference,
+  }) {
+    return CustomItemWidget(
+      nodeReference: nodeReference,
+    );
+  }
+}
+
+/*
+  A custom widget used to display data
+  reference the data from the node through the
+  nodeReference that can be passed through the CustomItem build
+ */
+class CustomItemWidget extends StatelessWidget {
+  final MAJNode nodeReference; // recommend always adding
+
+  const CustomItemWidget({
+    Key? key,
+    required this.nodeReference, // recommend always adding
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      children: [
+        // a back button to navigate back to the parent widget
+        ElevatedButton(
+          child: const Text("back"),
+          onPressed: () {
+            context.read<MAJProvider>().navigateToByNode(
+                  nodeReference.parent!,
+                );
+          },
+        ),
+        TextFormField(
+          initialValue: nodeReference.data!["name"], // get custom data
+          onChanged: (value) {
+            nodeReference.data!["name"] = value; // set custom data
+          },
+        ),
+      ],
+    );
+  }
+}
+
+void main() {
+  // add custom item's definition to MAJNode's static definitions
+  // important this is done before any custom items are built
+  MAJNode.addDefinition(
+    typeName: CustomItem.typeName,
+    function: () {
+      return CustomItem();
+    },
+  );
+
+  /*
+  // unsafe way to add a definition
+  MAJNode.definitions[CustomItem.typeName] = () {
+    return CustomItem();
+  };
+   */
+
+  // build an example tree
   MAJNode root = MAJNode(
     name: "the rooter",
-    typeName: typeName,
+    typeName: MAJDirectory.typeName,
     child: MAJDirectory(),
   );
 
   MAJNode one = root.addChild(MAJNode(
     name: "One",
-    typeName: typeName,
+    typeName: MAJDirectory.typeName,
     child: MAJDirectory(),
   ));
   MAJNode two = root.addChild(MAJNode(
     name: "Two",
-    typeName: typeName,
+    typeName: MAJDirectory.typeName,
     child: MAJDirectory(),
   ));
   MAJNode three = root.addChild(MAJNode(
     name: "Three",
-    typeName: typeName,
+    typeName: MAJDirectory.typeName,
     child: MAJDirectory(),
   ));
 
   one.addChild(MAJNode(
     name: "one 1",
-    typeName: typeName,
+    typeName: MAJDirectory.typeName,
     child: MAJDirectory(),
   ));
   one.addChild(MAJNode(
     name: "one 2",
-    typeName: typeName,
+    typeName: MAJDirectory.typeName,
     child: MAJDirectory(),
   ));
   one.addChild(MAJNode(
     name: "one 3",
-    typeName: typeName,
+    typeName: MAJDirectory.typeName,
     child: MAJDirectory(),
   ));
 
   MAJNode deep = two.addChild(MAJNode(
     name: "deep",
-    typeName: typeName,
+    typeName: MAJDirectory.typeName,
     child: MAJDirectory(),
   ));
   deep.addChild(MAJNode(
     name: "deep 1",
-    typeName: typeName,
+    typeName: MAJDirectory.typeName,
     child: MAJDirectory(),
   ));
   deep.addChild(MAJNode(
     name: "deep 2",
-    typeName: typeName,
+    typeName: MAJDirectory.typeName,
     child: MAJDirectory(),
   ));
 
   three.addChild(MAJNode(
     name: "three 1",
-    typeName: typeName,
+    typeName: MAJDirectory.typeName,
     child: MAJDirectory(),
   ));
 
   MAJNode bruh = MAJNode(
     name: "bruh",
-    typeName: typeName,
+    typeName: MAJDirectory.typeName,
     child: MAJDirectory(),
   );
   bruh.addChild(MAJNode(
     name: "Honey",
-    typeName: typeName,
+    typeName: MAJDirectory.typeName,
     child: MAJDirectory(),
   ));
-
   three.addChild(bruh);
 
+  // add a custom item to the root, and deep
+  root.addChild(
+    MAJNode(
+      name: "custom",
+      typeName: CustomItem.typeName,
+      data: <String, dynamic>{
+        "name": "Snarf",
+      },
+      child: CustomItem(),
+    ),
+  );
+
+  deep.addChild(
+    MAJNode(
+      name: "deep custom",
+      typeName: CustomItem.typeName,
+      data: {
+        "name": "narf",
+      },
+      child: CustomItem(),
+    ),
+  );
+
+  // print a breadth first traversal so the structure can be seen
   print(root.breadthFirstTraversal());
+
+  // convert to json as a proof then build from json
   MAJNode newer = MAJNode.fromJson(root.breadthFirstToJson());
 
   runApp(
     MaterialApp(
       home: Scaffold(
         body: SafeArea(
-          child: Center(
-            child: MAJBuilder(
-              root: newer,
-            ),
+          child: MAJBuilder(
+            root: newer,
           ),
         ),
       ),