+++ lib/src/maj_provider.dart
@@ -7,12 +7,16 @@
  */
 
 import 'package:flutter/material.dart';
+import 'package:memory_and_json_directories/memory_and_json_directories.dart';
 
 class MAJProvider with ChangeNotifier {
   String currentPath;
+  MAJNode currentNode;
+  bool byPathElseByNode = false;
 
   MAJProvider({
     required this.currentPath,
+    required this.currentNode,
   });
 
   /// pass a valid path ex /root/node
@@ -21,7 +25,14 @@ class MAJProvider with ChangeNotifier {
   /// ex:
   /// context.read<MAJProvider>().navigateTo("/root/node");
   void navigateTo(String path) {
+    byPathElseByNode = true;
     currentPath = path;
+    notifyListeners();
+  }
+
+  void navigateToByNode(MAJNode nodeTo) {
+    byPathElseByNode = false;
+    currentNode = nodeTo;
     notifyListeners();
   }
 }